Renamed entity and service files
